Finish level 04_overlay
Instead of copying all the image files into the directory, we need to
reuse the files. That's why we should use `overlay` filesystem. So for
the first time running the container, we should extract the image files
into a specific location (whatever you want). And then when we should
mount it to the lower directory. And anything we changed would be inside
the upper directory. And the merged directory is `rootfs`. What a
wonderful way!

Additionally, refactor the code.